global: adds autocomplete basis, combined lists
* Adds the basic idea to perform a quick search on certain
  values (circ_id_complete.js)
* Reworks list queries for improved performance.
  Some general queries seem to be quicker in sql.
* Adds a way to combine multiple circulation lists in one.

+++ invenio_circulation/views/lists.py
@@ -40,9 +40,57 @@ def lists_overview():
                            active_nav='lists', lists=lists)
 
 
+def _get_merged_headers(table_headers):
+    # TODO: at some point using an OrderedSet seems to be a cleaner idea
+    from collections import OrderedDict
+
+    def merge(dict1, dict2):
+        dict1.update(dict2)
+        return dict1
+
+    dicts = []
+    for table_header in table_headers:
+        tmp = OrderedDict()
+        for column in table_header:
+            tmp[column] = None
+        dicts.append(tmp)
+
+    merged_headers = reduce(merge, dicts, OrderedDict()).keys()
+
+    return merged_headers
+
+
+def _multiple_lists(list_links):
+    from invenio_circulation.signals import lists_class
+
+    clazzes = []
+    for list_link in list_links:
+        clazzes.append(send_signal(lists_class, list_link, None)[0])
+
+    table_header = _get_merged_headers(clazz.table_header for clazz in clazzes)
+    items = []
+    for clazz in clazzes:
+        for item in clazz.data():
+            tmp = []
+            for th in table_header:
+                tmp.append(item['item'][th] if th in item['item'] else '')
+            item['item'] = tmp
+            items.append(item)
+
+
+    modals = [('acquisition_vendor_price', 'Enter Vendor ID and price',
+               [('vendor_id', 'Vendor ID'), ('price', 'Price')])]
+    return render_template('lists/multiple_lists.html',
+                           active_nav='lists',
+                           table_header=table_header, items=items,
+                           modals=modals)
+
+
 @blueprint.route('/lists/<list_link>')
 @cap.require(403)
 def list_entrance(list_link):
+    if ',' in list_link:
+        return _multiple_lists(list_link.split(','))
     from invenio_circulation.signals import lists_class
 
     clazz = send_signal(lists_class, list_link, None)[0]
